Add CLI option to override shell used
As noted in #117, we want to override the shell that's used.

To do so, we can continue using the default of `bash`, but allow
overriding it with the `--shell` flag.

We need to disable the gosec finding for the new shell argument:

  G204: Subprocess launched with a potential tainted input or cmd arguments (gosec)

As we're comfortable with the risk, as it runs on the user's machine.

Closes #117.